refactor: extract MyTicketsTable component
